+++ Projects/Domain/Sources/Protocol/Repository/AuthRepositoryProtocol.swift
@@ -25,7 +25,4 @@ public protocol AuthRepositoryProtocol {
 
     /// 계정 탈퇴를 진행합니다.
     func withdraw() async throws
-
-    /// 토큰 재발급을 진행합니다.
-    func reissueToken() async throws
 }
